[m] fixed a test

+++ examples/calc.hs
@@ -20,7 +20,7 @@
 import Python5.Builtin
 
 main = do
-    print(float(1 / 2))
+    print(1 / 2)
     print(2 ** 3)
     print(17 / 3)
     print(17 // 3)
